feat: #15 icloud service 생성
icloud service로 Manager  로직 분리
